+++ app-dev-debian.4/.manala/debian/virtualbox/install_additions.sh
@@ -8,16 +8,18 @@ export DEBIAN_FRONTEND=noninteractive
 apt-get --yes purge --auto-remove virtualbox-\*
 
 # Install dependencies
-apt-get --quiet --verbose-versions --yes --no-install-recommends install bzip2
+apt-get --quiet --verbose-versions --yes --no-install-recommends install bzip2 dkms
 
-# Install build dependencies & packages
-apt-get --quiet --verbose-versions --yes --no-install-recommends install dkms linux-headers-$(uname -r)
+# Install build dependencies
+apt-get --quiet --verbose-versions --yes --no-install-recommends install linux-headers-amd64
 
 # Install the virtualbox guest additions
-mount --options loop ~/VBoxGuestAdditions.iso /mnt
-sh /mnt/VBoxLinuxAdditions.run
-umount /mnt
+mkdir -p /mnt/VBoxGuestAdditions
+mount --options loop ~/VBoxGuestAdditions.iso /mnt/VBoxGuestAdditions
+sh /mnt/VBoxGuestAdditions/VBoxLinuxAdditions.run
+umount /mnt/VBoxGuestAdditions
+rm -Rf /mnt/VBoxGuestAdditions
 rm -f VBoxGuestAdditions.iso
 
-# Remove build dependencies & packages
-apt-get --yes --auto-remove purge dkms linux-headers-$(uname -r)
+# Remove build dependencies
+apt-get --yes --auto-remove purge linux-headers-amd64